燕鸣书屋

+++ src/main/java/common/model/BookProfessionalList.java
@@ -5,7 +5,7 @@ public class BookProfessionalList {
 
     private String bookName;
 
-    private Integer isbn;
+    private Long isbn;
 
     private Integer typeProfessional1;
 
@@ -37,7 +37,7 @@ public class BookProfessionalList {
 
     private Integer isDelete;
 
-    public BookProfessionalList(Integer id, String bookName, Integer isbn, Integer typeProfessional1, String typeProfessionalname1, Integer typeProfessional2, String typeProfessionalname2, Integer grade, Integer semester, Integer printUser, Integer printTime, Integer importUser, Integer importTime, Integer exportUser, Integer exportTime, Integer cU, Integer cT, Integer isDelete) {
+    public BookProfessionalList(Integer id, String bookName, Long isbn, Integer typeProfessional1, String typeProfessionalname1, Integer typeProfessional2, String typeProfessionalname2, Integer grade, Integer semester, Integer printUser, Integer printTime, Integer importUser, Integer importTime, Integer exportUser, Integer exportTime, Integer cU, Integer cT, Integer isDelete) {
         this.id = id;
         this.bookName = bookName;
         this.isbn = isbn;
@@ -78,11 +78,11 @@ public class BookProfessionalList {
         this.bookName = bookName == null ? null : bookName.trim();
     }
 
-    public Integer getIsbn() {
+    public Long getIsbn() {
         return isbn;
     }
 
-    public void setIsbn(Integer isbn) {
+    public void setIsbn(Long isbn) {
         this.isbn = isbn;
     }
